+++ cmd/gateway-unsupported.go
@@ -16,28 +16,44 @@
 
 package cmd
 
+import "io"
+
+type gatewayUnsupported struct{}
+
+// CopyObjectPart - Not implemented.
+func (a gatewayUnsupported) CopyObjectPart(srcBucket, srcObject, destBucket, destObject string, uploadID string,
+	partID int, startOffset int64, length int64) (info PartInfo, err error) {
+	return info, traceError(NotImplemented{})
+}
+
+// AnonPutObject creates a new object anonymously with the incoming data,
+func (a gatewayUnsupported) AnonPutObject(bucket, object string, size int64, data io.Reader,
+	metadata map[string]string, sha256sum string) (ObjectInfo, error) {
+	return ObjectInfo{}, traceError(NotImplemented{})
+}
+
 // HealBucket - Not relevant.
-func (a *azureObjects) HealBucket(bucket string) error {
+func (a gatewayUnsupported) HealBucket(bucket string) error {
 	return traceError(NotImplemented{})
 }
 
 // ListBucketsHeal - Not relevant.
-func (a *azureObjects) ListBucketsHeal() (buckets []BucketInfo, err error) {
+func (a gatewayUnsupported) ListBucketsHeal() (buckets []BucketInfo, err error) {
 	return nil, traceError(NotImplemented{})
 }
 
 // HealObject - Not relevant.
-func (a *azureObjects) HealObject(bucket, object string) (int, int, error) {
+func (a gatewayUnsupported) HealObject(bucket, object string) (int, int, error) {
 	return 0, 0, traceError(NotImplemented{})
 }
 
 // ListObjectsHeal - Not relevant.
-func (a *azureObjects) ListObjectsHeal(bucket, prefix, marker, delimiter string, maxKeys int) (loi ListObjectsInfo, e error) {
+func (a gatewayUnsupported) ListObjectsHeal(bucket, prefix, marker, delimiter string, maxKeys int) (loi ListObjectsInfo, e error) {
 	return loi, traceError(NotImplemented{})
 }
 
 // ListUploadsHeal - Not relevant.
-func (a *azureObjects) ListUploadsHeal(bucket, prefix, marker, uploadIDMarker,
+func (a gatewayUnsupported) ListUploadsHeal(bucket, prefix, marker, uploadIDMarker,
 	delimiter string, maxUploads int) (lmi ListMultipartsInfo, e error) {
 	return lmi, traceError(NotImplemented{})
 }
