--- cmd/to_err_test.go
@@ -1,31 +1,0 @@
-/*
- * Minio Cloud Storage, (C) 2017 Minio, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cmd
-
-import "testing"
-
-func TestToErrIsNil(t *testing.T) {
-	if toObjectErr(nil) != nil {
-		t.Errorf("Test expected to return nil, failed instead got a non-nil value %s", toObjectErr(nil))
-	}
-	if toStorageErr(nil) != nil {
-		t.Errorf("Test expected to return nil, failed instead got a non-nil value %s", toStorageErr(nil))
-	}
-	if toAPIErrorCode(nil) != ErrNone {
-		t.Errorf("Test expected error code to be ErrNone, failed instead provided %d", toAPIErrorCode(nil))
-	}
-}
